Cleanup 1.5/1.6+ version detection
Removing the init-time version detection and using build tags to differentiate
between 1.5 and 1.6 (like the other versions).
